refactor: rename query traffic packages; refine ranking interface
