+++ integration/trafficcache/querytraffic/interface.go
@@ -1,4 +1,4 @@
-package inquirer
+package querytraffic
 
 import (
 	proxy "github.com/Telenav/osrm-backend/integration/pkg/trafficproxy"
